+++ app/Http/Controllers/LinkController.php
@@ -4,7 +4,6 @@ namespace App\Http\Controllers;
 
 use App\Link;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Storage;
 
 class LinkController extends Controller
 {
@@ -17,9 +16,6 @@ class LinkController extends Controller
     {
         return view('form-link')->with([
             'page_title' => __('Add link'),
-            'submit' => route('api.link.store'),
-            'parse' => route('api.link.parse'),
-            'method' => 'POST',
             'query' => $request->query('url')
         ]);
     }
@@ -30,30 +26,7 @@ class LinkController extends Controller
 
         return view('form-link')->with([
             'page_title' => __('Update link'),
-            'submit' => route('api.link.update', $id),
-            'parse' => route('api.link.parse'),
-            'method' => 'PUT',
             'link' => $link,
         ]);
-    }
-
-    public function delete(Request $request, int $id, string $hash)
-    {
-        if ($hash !== csrf_token()) {
-            abort(403);
-        }
-
-        /** @var Link $link */
-        $link = Link::with('post')->findOrFail($id);
-
-        if ($link->hasArchive()) {
-            Storage::disk('archives')->delete($link->archive);
-        }
-
-        $link->delete();
-        $link->post->delete();
-
-        $this->flash(__('Link :name has been deleted', ['name' => $link->title]), 'success');
-        return redirect()->back();
     }
 }