+++ routes/api.php
@@ -10,6 +10,8 @@ Route::get('tags', 'TagsController@all')->name('tags');
 Route::post('link/parse', 'LinkController@parse')->name('link.parse');
 Route::post('link', 'LinkController@store')->name('link.store');
 Route::put('link/{id}', 'LinkController@update')->name('link.update');
+Route::delete('link/{id}', 'LinkController@delete')->name('link.delete');
+Route::put('link/{id}/preview', 'LinkController@preview')->name('link.preview');
 
 Route::post('story', 'StoryController@store')->name('story.store');
 Route::put('story/{id}', 'StoryController@update')->name('story.update');