+++ routes/web.php
@@ -13,12 +13,11 @@ Route::get('/tag/{tag}', 'BrowseController@tag')->name('tag');
 
 Route::get('link/create', 'LinkController@create')->name('link.create');
 Route::get('link/edit/{id}', 'LinkController@edit')->name('link.edit');
-Route::get('link/delete/{id}/{hash}', 'LinkController@delete')->name('link.delete');
-Route::get('link/update-preview/{id}/{hash}', 'LinkActionsController@updatePreview')->name('link.update-preview');
-Route::get('link/archive/{id}', 'LinkActionsController@archiveForm')->name('link.archive-form');
-Route::post('link/archive/{id}', 'LinkActionsController@archiveStore');
-Route::post('link/delete-archive/{id}', 'LinkActionsController@archiveDelete')->name('link.archive-delete');
-Route::get('link/download-archive/{id}/{hash}', 'LinkActionsController@archiveDownload')->name('link.archive-download');
+
+Route::get('link/archive/{id}', 'LinkArchiveController@form')->name('link.archive.form');
+Route::post('link/archive/{id}', 'LinkArchiveController@store');
+Route::post('link/archive/{id}/delete', 'LinkArchiveController@delete')->name('link.archive.delete');
+Route::get('link/archive/{id}/download/{hash}', 'LinkArchiveController@download')->name('link.archive.download');
 
 Route::get('story/create', 'StoryController@create')->name('story.create');
 Route::get('story/{id}/edit', 'StoryController@edit')->name('story.edit');